Move Filterable into Functor module.
