+++ src/Kindly.hs
@@ -7,7 +7,6 @@ where
 
 import Kindly.Bifunctor as M
 import Kindly.Class as M
-import Kindly.Filterable as M
 import Kindly.Functor as M
 import Kindly.Iso as M
 import Kindly.Rank2 as M
